COZMO-15672: Simulated radio may silently drop packets (#6469)
* COZMO-15259: Simulated radio may silently drop packets

Utility class UdpServer.Send() returns (int) -1 to indicate send error.
In sim_radio.cpp, this return value is stored into u32, so it is treated as a large positive value.
This causes the check for "if (sent_bytes < length)" to succeed even though data was not sent.

Dropped packets may be contributing to unexplained unit test failures.

Fix is to store return value as int and perform signed comparison when checking for error.

If udp server is dropping packets, webots process will log error and drop the connection.

* COZMO-15259: Bonus fix - typos

* Auto committing firmware from dev/dmudie/COZMO-15259-radio-drops-messages

* COZMO-15672: Fix message send/recv functions to check for send/recv errors

@kevinatanki pointed out that signed result values were being mishandled in several other places.

I went through various layers of message transport (messaging/basestation, messaging/shared, GameComms, UdpSocketComms, MultiClientComms etc) and changed send/recv functions to return ssize_t instead of size_t, u32, or int. I then changed counters and result checks to match.

I went with ssize_t because it can represent -1 as well as the maximum size returned by any calls to send or recv.

I noticed a number of places where send/recv errors were checked but not logged.  Shouldn't failed send/recv be logged for diagnostics?

* Auto committing firmware from dev/dmudie/COZMO-15259-radio-drops-messages

+++ coretech/messaging/include/anki/messaging/shared/UdpServer.h
@@ -23,8 +23,8 @@ public:
   
   int GetNumClients();
 
-  int Send(const char* data, int size);
-  int Recv(char* data, int maxSize);
+  ssize_t Send(const char* data, int size);
+  ssize_t Recv(char* data, int maxSize);
   //int GetNumBytesAvailable();
   
 private:
